Move `update()` to Templater base class
